Refactor and document the code
