+++ packages/application/package.json
@@ -50,6 +50,7 @@
     "@jupyterlab/rendermime": "^3.1.0-alpha.11",
     "@jupyterlab/rendermime-interfaces": "^3.1.0-alpha.11",
     "@jupyterlab/services": "^6.1.0-alpha.11",
+    "@jupyterlab/settingregistry": "^3.1.0-alpha.11",
     "@jupyterlab/statedb": "^3.1.0-alpha.11",
     "@jupyterlab/translation": "^3.1.0-alpha.11",
     "@jupyterlab/ui-components": "^3.1.0-alpha.11",
