+++ packages/application/src/index.ts
@@ -17,6 +17,8 @@ export { JupyterLab } from './lab';
 
 export { ILayoutRestorer, LayoutRestorer } from './layoutrestorer';
 
+export { MenuFactory } from './menufactory';
+
 export {
   IMimeDocumentTracker,
   createRendermimePlugin,
